refactor: refactor SDK into smaller files
