+++ test/SgidClient.spec.ts
@@ -1,11 +1,16 @@
 import { readFileSync } from 'fs'
 
-import SgidClient from '../src'
 import {
+  generateCodeChallenge,
+  generateCodeVerifier,
+  generatePkcePair,
+  SgidClient,
+} from '../src'
+import {
+  codeVerifierAndChallengePattern,
   DEFAULT_SCOPE,
   DEFAULT_SGID_CODE_CHALLENGE_METHOD,
 } from '../src/constants'
-import { codeVerifierAndChallengePattern } from '../src/util'
 
 import {
   MOCK_ACCESS_TOKEN,
@@ -395,14 +400,14 @@ describe('SgidClient', () => {
 
   describe('generatePkcePair', () => {
     it('can generate a PKCE pair', () => {
-      const pkcePair = SgidClient.generatePkcePair()
+      const pkcePair = generatePkcePair()
       expect(pkcePair).toBeDefined()
     })
   })
 
   describe('generateCodeVerifier', () => {
     it('should generate a code verifier of length 43 when no length is provided', () => {
-      const codeVerifier = SgidClient.generateCodeVerifier()
+      const codeVerifier = generateCodeVerifier()
 
       expect(codeVerifier.length).toBe(43)
       expect(codeVerifier).toMatch(codeVerifierAndChallengePattern)
@@ -410,7 +415,7 @@ describe('SgidClient', () => {
 
     it('should generate a code verifier of specified length when length between 43 (inclusive) and 128 (inclusive) is provided', () => {
       for (let length = 43; length <= 128; length++) {
-        const codeVerifier = SgidClient.generateCodeVerifier(length)
+        const codeVerifier = generateCodeVerifier(length)
         expect(codeVerifier.length).toBe(length)
         expect(codeVerifier).toMatch(codeVerifierAndChallengePattern)
       }
@@ -418,7 +423,7 @@ describe('SgidClient', () => {
 
     it('should throw an error when a length < 43 or length > 128 is provided', () => {
       for (const length of [-1, 0, 42, 129, 138, 999]) {
-        expect(() => SgidClient.generateCodeVerifier(length)).toThrowError(
+        expect(() => generateCodeVerifier(length)).toThrowError(
           `The code verifier should have a minimum length of 43 and a maximum length of 128. Length of ${length} was provided`,
         )
       }
@@ -427,16 +432,14 @@ describe('SgidClient', () => {
 
   describe('generateCodeChallenge', () => {
     it('should match the specified pattern', () => {
-      expect(SgidClient.generateCodeChallenge(MOCK_CODE_VERIFIER)).toMatch(
+      expect(generateCodeChallenge(MOCK_CODE_VERIFIER)).toMatch(
         codeVerifierAndChallengePattern,
       )
     })
 
     it('should be deterministic (return the same code challenge given the same code verifier)', () => {
-      const firstCodeChallenge =
-        SgidClient.generateCodeChallenge(MOCK_CODE_VERIFIER)
-      const secondCodeChallenge =
-        SgidClient.generateCodeChallenge(MOCK_CODE_VERIFIER)
+      const firstCodeChallenge = generateCodeChallenge(MOCK_CODE_VERIFIER)
+      const secondCodeChallenge = generateCodeChallenge(MOCK_CODE_VERIFIER)
 
       expect(firstCodeChallenge).toBe(secondCodeChallenge)
     })
